Part 2: add sharedCache follow leveldb

+++ lru/lru.go
@@ -5,8 +5,8 @@ import "container/list"
 // todo 添加LFU和LRU-K
 
 type Cache struct {
-	maxBytes  int64
-	nBytes    int64
+	maxBytes  uint32
+	nBytes    uint32
 	cache     map[string]*list.Element
 	ll        *list.List
 	OnEvicted func(key string, value Value)
@@ -21,7 +21,7 @@ type Value interface {
 	Len() int
 }
 
-func NewCache(maxBytes int64, onEvicted func(key string, value Value)) *Cache {
+func NewCache(maxBytes uint32, onEvicted func(key string, value Value)) *Cache {
 	return &Cache{
 		maxBytes:  maxBytes,
 		ll:        list.New(),
@@ -45,7 +45,7 @@ func (c *Cache) RemoveOldest() {
 		c.ll.Remove(ele)
 		kv := ele.Value.(*entry)
 		delete(c.cache, kv.key)
-		c.nBytes -= int64(len(kv.key)) + int64(kv.val.Len())
+		c.nBytes -= uint32(len(kv.key)) + uint32(kv.val.Len())
 		if c.OnEvicted != nil {
 			c.OnEvicted(kv.key, kv.val)
 		}
@@ -57,11 +57,11 @@ func (c *Cache) Add(key string, value Value) {
 		c.ll.MoveToFront(ele)
 		kv := ele.Value.(*entry)
 		kv.val = value
-		c.nBytes += int64(kv.val.Len()) - int64(value.Len())
+		c.nBytes += uint32(kv.val.Len()) - uint32(value.Len())
 	} else {
 		ele = c.ll.PushFront(&entry{key: key, val: value})
 		c.cache[key] = ele
-		c.nBytes += int64(len(key)) + int64(value.Len())
+		c.nBytes += uint32(len(key)) + uint32(value.Len())
 	}
 	for c.maxBytes != 0 && c.nBytes > c.maxBytes {
 		c.RemoveOldest()
